authorization changes

+++ src/services/talukaService.js
@@ -1,6 +1,11 @@
 // src/services/talukaService.js
 
 import Taluka from "../models/talukaModel.js";
+import Village from "../models/villageModel.js";
+import Complainer from "../models/complainerModel.js";
+import Complaint from "../models/complaintModel.js";
+import Admin from "../models/adminModel.js";
+import Visitor from "../models/visitorModel.js";
 import Counter from "../models/counterModel.js";
 import { generateTalukaId } from "../utils/generateIds.js";
 
@@ -42,7 +47,7 @@ export const updateTalukaService = async (talukaId, name) => {
         mr: name.mr
       }
     },
-    { new: true }
+    { new: true, runValidators: true }
   );
 
   if (!updated) {
@@ -54,6 +59,39 @@ export const updateTalukaService = async (talukaId, name) => {
 
 /* ================= DELETE TALUKA ================= */
 export const deleteTalukaService = async (talukaId) => {
+  const taluka = await Taluka.findOne({ talukaId }).select("_id");
+  if (!taluka) {
+    throw new Error("Taluka not found");
+  }
+
+  const [villageCount, complainerCount, adminCount, visitorCount] = await Promise.all([
+    Village.countDocuments({ taluka: taluka._id }),
+    Complainer.countDocuments({ taluka: taluka._id }),
+    Admin.countDocuments({ assignedTaluka: taluka._id }),
+    Visitor.countDocuments({ taluka: taluka._id })
+  ]);
+
+  let complaintCount = 0;
+  if (complainerCount > 0) {
+    const complainerIds = await Complainer.find(
+      { taluka: taluka._id },
+      { _id: 1 }
+    ).lean();
+    complaintCount = await Complaint.countDocuments({
+      complainer: { $in: complainerIds.map((c) => c._id) }
+    });
+  }
+
+  if (
+    villageCount > 0 ||
+    complainerCount > 0 ||
+    complaintCount > 0 ||
+    adminCount > 0 ||
+    visitorCount > 0
+  ) {
+    throw new Error("Cannot delete taluka because dependent records exist");
+  }
+
   const deleted = await Taluka.findOneAndDelete({ talukaId });
 
   if (!deleted) {
@@ -65,6 +103,11 @@ export const deleteTalukaService = async (talukaId) => {
 
 /* ================= RESET TALUKA COUNTER ================= */
 export const resetTalukaCounterService = async () => {
+  const talukaCount = await Taluka.countDocuments();
+  if (talukaCount > 0) {
+    throw new Error("Cannot reset taluka counter while taluka records exist");
+  }
+
   await Counter.findByIdAndUpdate(
     "talukaId",
     { seq: 0 },
